test(commands): consolidate test files from 49 to 33 files
Merge duplicate test files while preserving all 279 test cases:
- Handoff: 5 → 2 files (commands + helpers)
- Inspect: 6 → 2 files (cli + helpers)
- PR: 5 → 2 files (ready + show)
- Flow: 2 → 1 file (update)
- Others: 5 → 3 files (output_format, check, task_binding)
- Removed commands: 1 → 1 file

Enhanced conftest.py with shared fixtures:
- cli_runner: eliminate 28 duplicate CliRunner instances
- mock_flow_service, mock_pr_service: standard service mocks
- mock_git_client: git operations mock
- mock_handoff_service: handoff operations mock
- mock_flow_status_response: flow status model factory
- inspect_app_fixture: inspect app import

All 279 tests pass. No test logic changed.
