--- tests/vibe3/commands/test_inspect_commands.py
@@ -1,42 +1,0 @@
-"""Tests for vibe inspect commands subcommand.
-
-Tests CLI surface: argument validation, help output, exit codes.
-All external services are mocked.
-"""
-
-from unittest.mock import MagicMock, patch
-
-from typer.testing import CliRunner
-
-from vibe3.commands.inspect import app
-
-runner = CliRunner()
-
-
-def test_inspect_commands_no_args_lists_available():
-    """commands 不传参数时列出可分析的顶层命令，不报错。"""
-    result = runner.invoke(app, ["commands"])
-    assert result.exit_code == 0
-    assert "=== vibe3 command structure ===" in result.output
-    assert "Top-level commands:" in result.output
-    assert "flow" in result.output
-    assert "pr" in result.output
-    assert "review" in result.output
-    assert "inspect" in result.output
-    assert "symbols" not in result.output
-    assert "files" not in result.output
-
-
-def test_inspect_commands_with_command():
-    mock_result = MagicMock()
-    mock_result.command = "review"
-    mock_result.file_path = "src/vibe3/commands/review.py"
-    mock_result.call_depth = 2
-    mock_result.calls = []
-    mock_result.model_dump.return_value = {}
-    with patch(
-        "vibe3.commands.inspect.command_analyzer.analyze_command",
-        return_value=mock_result,
-    ):
-        result = runner.invoke(app, ["commands", "review"])
-    assert result.exit_code == 0
